% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{arrow_array}
\alias{arrow_array}
\alias{as_arrow_array}
\alias{as_arrow_array.arrowvctrs_array}
\alias{as_arrow_buffer}
\alias{as_arrow_buffer.default}
\title{Create Arrow arrays}
\usage{
arrow_array(
  buffers = NULL,
  length = 0,
  null_count = -1,
  offset = 0,
  children = NULL,
  dictionary = NULL
)

as_arrow_array(x, ...)

\method{as_arrow_array}{arrowvctrs_array}(x, ...)

as_arrow_buffer(x, ...)

\method{as_arrow_buffer}{default}(x, ...)
}
\arguments{
\item{buffers}{A \code{\link[=list]{list()}} of R vectors whose memory serves as an underlying
buffer in the Arrow format.}

\item{length}{The logical size of the array}

\item{null_count}{The number of NULL values or -1 if this hasn't been
computed yet.}

\item{offset}{The number of elements to skip at the front of the array.}

\item{children}{Child vectors as a \code{\link[=list]{list()}} coerced by \code{\link[=as_arrow_array]{as_arrow_array()}}}

\item{dictionary}{Dictionary array for dictionary types.}

\item{x}{An object to convert to an arrow_array}

\item{...}{Passed to S3 Methods}
}
\value{
An object of class "arrowvctrs_array"
}
\description{
Create Arrow arrays
}
\examples{
arrow_array(1:100, 100)

}
