% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{arrow_deep_copy}
\alias{arrow_deep_copy}
\title{Create copies of Arrow vectors}
\usage{
arrow_deep_copy(x)
}
\arguments{
\item{x}{An \code{\link[=arrow_vctr]{arrow_vctr()}}.}
}
\value{
An \code{\link[=arrow_vctr]{arrow_vctr()}}
}
\description{
Create copies of Arrow vectors
}
\examples{
arrow_deep_copy(as_arrow_vctr(1:5))

}
