% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{arrow_schema}
\alias{arrow_schema}
\title{Create Arrow Schema objects}
\usage{
arrow_schema(
  format,
  name = NULL,
  metadata = NULL,
  flags = 0L,
  children = NULL,
  dictionary = NULL
)
}
\arguments{
\item{format}{A
\href{https://arrow.apache.org/docs/format/CDataInterface.html#data-type-description-format-strings}{format string}.}

\item{name}{An optional name}

\item{metadata}{A named character vector of \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}}
containing key/value metadata.}

\item{flags}{Flags to set on input.}

\item{children}{A \code{\link[=list]{list()}} of objects created using \code{\link[=arrow_schema]{arrow_schema()}}.}

\item{dictionary}{An \code{\link[=arrow_schema]{arrow_schema()}} if this is a dictionary type.}
}
\value{
An external pointer with class 'arrowvctrs_schema'
}
\description{
Create Arrow Schema objects
}
