% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{arrow_vctr}
\alias{arrow_vctr}
\title{Create arrow vectors}
\usage{
arrow_vctr(schema = arrow_schema("n"), arrays = list())
}
\arguments{
\item{schema}{An \code{\link[=arrow_schema]{arrow_schema()}}}

\item{arrays}{An \code{\link[=arrow_array]{arrow_array()}} or \code{\link[=list]{list()}} of \code{\link[=arrow_array]{arrow_array()}}s.}
}
\value{
An object of class "arrowvctrs_vctr"
}
\description{
Create arrow vectors
}
\examples{
arrow_vctr()

}
