% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset.R
\name{as_arrow_offset}
\alias{as_arrow_offset}
\title{Offset vectors}
\usage{
as_arrow_offset(x, ...)
}
\arguments{
\item{x}{An object to convert to an offset vector}

\item{...}{Passed to S3 Methods}
}
\value{
An object of class "arrowc_offset"
}
\description{
Offset vectors
}
\examples{
as_arrow_offset(1:5)

}
