% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr-base.R
\name{as_arrow_vctr.NULL}
\alias{as_arrow_vctr.NULL}
\alias{as_arrow_vctr.logical}
\alias{as_arrow_vctr.integer}
\alias{as_arrow_vctr.double}
\alias{as_arrow_vctr.raw}
\alias{as_arrow_vctr.data.frame}
\title{Convert base R objects to Arrow vectors}
\usage{
\method{as_arrow_vctr}{`NULL`}(x, ..., name = NULL)

\method{as_arrow_vctr}{logical}(x, ..., name = NULL)

\method{as_arrow_vctr}{integer}(x, ..., name = NULL)

\method{as_arrow_vctr}{double}(x, ..., name = NULL)

\method{as_arrow_vctr}{raw}(x, ..., name = NULL)

\method{as_arrow_vctr}{data.frame}(x, ..., name = NULL)
}
\arguments{
\item{x}{An object to convert to an \code{\link[=arrow_vctr]{arrow_vctr()}}}

\item{...}{Passed to S3 methods}

\item{name}{An optional name}
}
\value{
An \code{\link[=arrow_vctr]{arrow_vctr()}}
}
\description{
These methods return an \code{\link[=arrow_vctr]{arrow_vctr()}} for R objects that don't involve
copying or unnecessary allocating (with the exception of ALTREP objects,
which will be expanded).
}
\examples{
as_arrow_vctr(NULL)
as_arrow_vctr(c(TRUE, FALSE, NA))
as_arrow_vctr(1:10)
as_arrow_vctr(c(1.1, 2.2))
as_arrow_vctr(as.raw(0x00))
as_arrow_vctr(data.frame(x = 1:10, y = as.raw(1:10)))

}
