% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmask.R
\name{as_bitmask}
\alias{as_bitmask}
\alias{as_bitmask.default}
\alias{as_bitmask.arrowc_bitmask}
\alias{as_bitmask.logical}
\alias{new_bitmask}
\alias{as.logical.arrowc_bitmask}
\title{Create bitmask buffers}
\usage{
as_bitmask(x, ...)

\method{as_bitmask}{default}(x, ...)

\method{as_bitmask}{arrowc_bitmask}(x, ...)

\method{as_bitmask}{logical}(x, ...)

new_bitmask(buffer)

\method{as.logical}{arrowc_bitmask}(x, ...)
}
\arguments{
\item{x}{An object to convert to a bitmask}

\item{...}{Passed to S3 methods}

\item{buffer}{A \code{\link[=raw]{raw()}} vector containing the bitmask data}
}
\value{
An object with class 'arrowc_bitmask'
}
\description{
Create bitmask buffers
}
\examples{
as_bitmask(c(TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE))
as.logical(new_bitmask(as.raw(0x01)))

}
