% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-array.R
\name{as_carrow_array.NULL}
\alias{as_carrow_array.NULL}
\alias{as_carrow_array.logical}
\alias{as_carrow_array.integer}
\alias{as_carrow_array.double}
\alias{as_carrow_array.character}
\alias{as_carrow_array.factor}
\alias{as_carrow_array.raw}
\alias{as_carrow_array.data.frame}
\title{Convert R objects to Arrow vectors}
\usage{
\method{as_carrow_array}{`NULL`}(x, ..., name = NULL)

\method{as_carrow_array}{logical}(x, ..., name = NULL)

\method{as_carrow_array}{integer}(x, ..., name = NULL)

\method{as_carrow_array}{double}(x, ..., name = NULL)

\method{as_carrow_array}{character}(x, ..., name = NULL)

\method{as_carrow_array}{factor}(x, ..., name = NULL)

\method{as_carrow_array}{raw}(x, ..., name = NULL)

\method{as_carrow_array}{data.frame}(x, ..., name = NULL)
}
\arguments{
\item{x}{An object to convert to an \code{\link[=carrow_array]{carrow_array()}}}

\item{...}{Passed to S3 methods}

\item{name}{An optional name}
}
\value{
An \code{\link[=carrow_array]{carrow_array()}}
}
\description{
These methods return an \code{\link[=carrow_array]{carrow_array()}} for R objects that don't involve
copying or unnecessary allocating. Two excpetions are (1) ALTREP objects,
which will be expanded, and (2) character vectors, which will be converted
to UTF-8 and serialized as a single \code{\link[=raw]{raw()}} vector.
}
\examples{
as_carrow_array(NULL)
as_carrow_array(c(TRUE, FALSE, NA))
as_carrow_array(1:10)
as_carrow_array(c(1.1, 2.2))
as_carrow_array(as.raw(0x00))
as_carrow_array("fish")
as_carrow_array(data.frame(x = 1:10, y = as.raw(1:10)))

}
