% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int64.R
\name{as_carrow_int64}
\alias{as_carrow_int64}
\title{64-bit integer vectors}
\usage{
as_carrow_int64(x, ...)
}
\arguments{
\item{x}{An object to convert to a large integer vector}

\item{...}{Passed to S3 Methods}
}
\value{
An object of class "carrow_int64"
}
\description{
64-bit integer vectors
}
\examples{
as_carrow_int64(1:5)

}
