% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{carrow_array}
\alias{carrow_array}
\alias{carrow_array_validate}
\alias{as_carrow_array}
\alias{as_carrow_array.carrow_array}
\title{Create arrow vectors}
\usage{
carrow_array(
  schema = carrow_schema("n"),
  array_data = carrow_array_data(),
  validate = TRUE
)

carrow_array_validate(x)

as_carrow_array(x, ...)

\method{as_carrow_array}{carrow_array}(x, ...)
}
\arguments{
\item{schema}{An \code{\link[=carrow_schema]{carrow_schema()}}}

\item{array_data}{An \code{\link[=carrow_array_data]{carrow_array_data()}}}

\item{validate}{Use \code{FALSE} to skip validation}

\item{x}{An object to convert to an \code{\link[=carrow_array]{carrow_array()}}}

\item{...}{Passed to S3 methods}
}
\value{
An object of class "carrow_array"
}
\description{
Create arrow vectors
}
\examples{
carrow_array()

}
