% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-data.R
\name{carrow_array_data}
\alias{carrow_array_data}
\alias{carrow_array_data_info}
\alias{as_carrow_array_data}
\alias{as_carrow_array_data.arrowvctrs_array_data}
\alias{as_carrow_buffer}
\alias{as_carrow_buffer.default}
\title{Create Arrow arrays}
\usage{
carrow_array_data(
  buffers = NULL,
  length = 0,
  null_count = -1,
  offset = 0,
  children = NULL,
  dictionary = NULL
)

carrow_array_data_info(x, ...)

as_carrow_array_data(x, ...)

\method{as_carrow_array_data}{arrowvctrs_array_data}(x, ...)

as_carrow_buffer(x, ...)

\method{as_carrow_buffer}{default}(x, ...)
}
\arguments{
\item{buffers}{A \code{\link[=list]{list()}} of R vectors whose memory serves as an underlying
buffer in the Arrow format.}

\item{length}{The logical size of the array}

\item{null_count}{The number of NULL values or -1 if this hasn't been
computed yet.}

\item{offset}{The number of elements to skip at the front of the array.}

\item{children}{Child vectors as a \code{\link[=list]{list()}} coerced by \code{\link[=as_carrow_array_data]{as_carrow_array_data()}}}

\item{dictionary}{Dictionary array for dictionary types.}

\item{x}{An object to convert to an carrow_array}

\item{...}{Passed to S3 Methods}
}
\value{
An object of class "arrowvctrs_array_data"
}
\description{
Create Arrow arrays
}
\examples{
carrow_array_data(1:100, 100)

}
