% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-stream-function.R
\name{carrow_array_stream_function}
\alias{carrow_array_stream_function}
\title{Array streams from functions}
\usage{
carrow_array_stream_function(schema, fun)
}
\arguments{
\item{schema}{A \code{\link[=carrow_schema]{carrow_schema()}}}

\item{fun}{An R function whose return value is converted to
a \code{\link[=carrow_array]{carrow_array()}} and passed to the caller.}
}
\value{
An object of class "carrow_array_stream"
}
\description{
Array streams from functions
}
