% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{carrow_array_stream_get_schema}
\alias{carrow_array_stream_get_schema}
\alias{carrow_array_stream_get_next}
\alias{as_carrow_array_stream}
\alias{as_carrow_array_stream.carrow_stream}
\alias{as_carrow_array_stream.carrow_array}
\title{Stream objects as Arrow arrays}
\usage{
carrow_array_stream_get_schema(array_stream)

carrow_array_stream_get_next(array_stream)

as_carrow_array_stream(x, ..., chunk_size = NULL)

\method{as_carrow_array_stream}{carrow_stream}(x, ..., chunk_size = NULL)

\method{as_carrow_array_stream}{carrow_array}(x, ..., chunk_size = NULL)
}
\arguments{
\item{array_stream}{An object of class "carrow_stream"}

\item{x}{An R object to convert to an Arrow Stream}

\item{...}{Passed to S3 methods}

\item{chunk_size}{The number of elements to return in each chunk
or \code{NULL}.}
}
\value{
An object of class "carrow_stream"
}
\description{
Stream objects as Arrow arrays
}
