% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-ptype.R
\name{carrow_default_ptype}
\alias{carrow_default_ptype}
\title{Guess prototypes for Arrow schemas}
\usage{
carrow_default_ptype(schema)
}
\arguments{
\item{schema}{An \code{\link[=carrow_schema]{carrow_schema()}}}
}
\value{
A zero-length array prototype
}
\description{
Guess prototypes for Arrow schemas
}
\examples{
carrow_default_ptype(carrow_schema("n"))
carrow_default_ptype(
  carrow_schema("+s", children = list(carrow_schema("i", "colname")))
)

}
