% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{carrow_vctr}
\alias{carrow_vctr}
\alias{as_carrow_vctr}
\title{Create R vector wrappers around 'Arrow' arrays}
\usage{
carrow_vctr(array = carrow_array())

as_carrow_vctr(x, ...)
}
\arguments{
\item{array}{A \code{\link[=carrow_array]{carrow_array()}}}
}
\value{
An object of class 'carrow_vctr'
}
\description{
Create R vector wrappers around 'Arrow' arrays
}
\examples{
carrow_vctr(1:10)

}
