% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-array.R
\name{from_carrow_array}
\alias{from_carrow_array}
\alias{from_carrow_array.default}
\alias{from_carrow_array.NULL}
\alias{from_carrow_array.logical}
\alias{from_carrow_array.integer}
\alias{from_carrow_array.double}
\alias{from_carrow_array.raw}
\alias{from_carrow_array.character}
\alias{from_carrow_array.factor}
\alias{from_carrow_array.data.frame}
\title{Convert Arrow vectors to R objects}
\usage{
from_carrow_array(x, ptype = carrow_default_ptype(x$schema), ...)

\method{from_carrow_array}{default}(x, ptype = carrow_default_ptype(x$schema), ...)

\method{from_carrow_array}{`NULL`}(x, ptype, ...)

\method{from_carrow_array}{logical}(x, ptype, ...)

\method{from_carrow_array}{integer}(x, ptype, ...)

\method{from_carrow_array}{double}(x, ptype, ...)

\method{from_carrow_array}{raw}(x, ptype, ...)

\method{from_carrow_array}{character}(x, ptype, ...)

\method{from_carrow_array}{factor}(x, ptype, ...)

\method{from_carrow_array}{data.frame}(x, ptype, ...)
}
\arguments{
\item{x}{An \code{\link[=carrow_array]{carrow_array()}}}

\item{ptype}{An R object to use as a prototype}

\item{...}{Passed to S3 methods}
}
\value{
An object with the same class as \code{ptype}
}
\description{
Note that \code{\link[=from_carrow_array]{from_carrow_array()}} dispatches on \code{ptype}
}
\examples{
from_carrow_array(as_carrow_array(c(TRUE, FALSE, NA)), logical())

}
