% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-array.R
\name{from_narrow_array}
\alias{from_narrow_array}
\alias{from_narrow_array.default}
\alias{from_narrow_array.NULL}
\alias{from_narrow_array.logical}
\alias{from_narrow_array.integer}
\alias{from_narrow_array.double}
\alias{from_narrow_array.raw}
\alias{from_narrow_array.character}
\alias{from_narrow_array.factor}
\alias{from_narrow_array.data.frame}
\title{Convert Arrow vectors to R objects}
\usage{
from_narrow_array(x, ptype = narrow_default_ptype(x$schema), ...)

\method{from_narrow_array}{default}(x, ptype = narrow_default_ptype(x$schema), ...)

\method{from_narrow_array}{`NULL`}(x, ptype = narrow_default_ptype(x$schema), ...)

\method{from_narrow_array}{logical}(x, ptype = narrow_default_ptype(x$schema), ...)

\method{from_narrow_array}{integer}(x, ptype = narrow_default_ptype(x$schema), ...)

\method{from_narrow_array}{double}(x, ptype = narrow_default_ptype(x$schema), ...)

\method{from_narrow_array}{raw}(x, ptype = narrow_default_ptype(x$schema), ...)

\method{from_narrow_array}{character}(x, ptype = narrow_default_ptype(x$schema), ...)

\method{from_narrow_array}{factor}(x, ptype = narrow_default_ptype(x$schema), ...)

\method{from_narrow_array}{data.frame}(x, ptype = narrow_default_ptype(x$schema), ...)
}
\arguments{
\item{x}{An \code{\link[=narrow_array]{narrow_array()}}}

\item{ptype}{An R object to use as a prototype}

\item{...}{Passed to S3 methods}
}
\value{
An object with the same class as \code{ptype}
}
\description{
Note that \code{\link[=from_narrow_array]{from_narrow_array()}} dispatches on \code{ptype}
}
\examples{
from_narrow_array(as_narrow_array(c(TRUE, FALSE, NA)), logical())

}
