% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-array.R
\name{from_sparrow_array}
\alias{from_sparrow_array}
\alias{from_sparrow_array.default}
\alias{from_sparrow_array.NULL}
\alias{from_sparrow_array.logical}
\alias{from_sparrow_array.integer}
\alias{from_sparrow_array.double}
\alias{from_sparrow_array.raw}
\alias{from_sparrow_array.character}
\alias{from_sparrow_array.factor}
\alias{from_sparrow_array.data.frame}
\title{Convert Arrow vectors to R objects}
\usage{
from_sparrow_array(x, ptype = sparrow_default_ptype(x$schema), ...)

\method{from_sparrow_array}{default}(x, ptype = sparrow_default_ptype(x$schema), ...)

\method{from_sparrow_array}{`NULL`}(x, ptype = sparrow_default_ptype(x$schema), ...)

\method{from_sparrow_array}{logical}(x, ptype = sparrow_default_ptype(x$schema), ...)

\method{from_sparrow_array}{integer}(x, ptype = sparrow_default_ptype(x$schema), ...)

\method{from_sparrow_array}{double}(x, ptype = sparrow_default_ptype(x$schema), ...)

\method{from_sparrow_array}{raw}(x, ptype = sparrow_default_ptype(x$schema), ...)

\method{from_sparrow_array}{character}(x, ptype = sparrow_default_ptype(x$schema), ...)

\method{from_sparrow_array}{factor}(x, ptype = sparrow_default_ptype(x$schema), ...)

\method{from_sparrow_array}{data.frame}(x, ptype = sparrow_default_ptype(x$schema), ...)
}
\arguments{
\item{x}{An \code{\link[=sparrow_array]{sparrow_array()}}}

\item{ptype}{An R object to use as a prototype}

\item{...}{Passed to S3 methods}
}
\value{
An object with the same class as \code{ptype}
}
\description{
Note that \code{\link[=from_sparrow_array]{from_sparrow_array()}} dispatches on \code{ptype}
}
\examples{
from_sparrow_array(as_sparrow_array(c(TRUE, FALSE, NA)), logical())

}
