% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{narrow_array}
\alias{narrow_array}
\alias{narrow_array_validate}
\alias{as_narrow_array}
\alias{as_narrow_array.narrow_array}
\title{Create arrow vectors}
\usage{
narrow_array(
  schema = narrow_schema("n"),
  array_data = narrow_array_data(),
  validate = TRUE
)

narrow_array_validate(x)

as_narrow_array(x, ...)

\method{as_narrow_array}{narrow_array}(x, ...)
}
\arguments{
\item{schema}{An \code{\link[=narrow_schema]{narrow_schema()}}}

\item{array_data}{An \code{\link[=narrow_array_data]{narrow_array_data()}}}

\item{validate}{Use \code{FALSE} to skip validation}

\item{x}{An object to convert to an \code{\link[=narrow_array]{narrow_array()}}}

\item{...}{Passed to S3 methods}
}
\value{
An object of class "narrow_array"
}
\description{
Create arrow vectors
}
\examples{
narrow_array()

}
