% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-stream.R
\name{narrow_array_stream}
\alias{narrow_array_stream}
\alias{narrow_array_stream_collect}
\alias{narrow_array_stream_get_schema}
\alias{narrow_array_stream_get_next}
\alias{as_narrow_array_stream}
\alias{as_narrow_array_stream.narrow_array_stream}
\alias{as_narrow_array_stream.list}
\alias{as_narrow_array_stream.function}
\alias{as_narrow_array_stream.default}
\alias{as_narrow_array_stream.narrow_array}
\title{Stream objects as Arrow arrays}
\usage{
narrow_array_stream(list_of_array = list(), schema = NULL, validate = TRUE)

narrow_array_stream_collect(array_stream, ptype = NULL)

narrow_array_stream_get_schema(array_stream)

narrow_array_stream_get_next(array_stream, validate = TRUE)

as_narrow_array_stream(x, ...)

\method{as_narrow_array_stream}{narrow_array_stream}(x, ...)

\method{as_narrow_array_stream}{list}(x, ...)

\method{as_narrow_array_stream}{`function`}(x, ...)

\method{as_narrow_array_stream}{default}(x, ...)

\method{as_narrow_array_stream}{narrow_array}(x, ...)
}
\arguments{
\item{list_of_array}{A \code{list()} of \code{\link[=narrow_array]{narrow_array()}}s.}

\item{schema}{The schema for all the arrays in \code{list_of_array}.}

\item{validate}{Use \code{FALSE} to skip validation of arrays in
\code{list_of_array}.}

\item{array_stream}{An object of class "narrow_array_stream"}

\item{ptype}{An R object to use as a prototype}

\item{x}{An R object to convert to an Arrow Stream}

\item{...}{Passed to S3 methods}
}
\value{
An object of class "narrow_array_stream"
}
\description{
Stream objects as Arrow arrays
}
