% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{narrow_deep_copy}
\alias{narrow_deep_copy}
\title{Create copies of Arrow vectors}
\usage{
narrow_deep_copy(x)
}
\arguments{
\item{x}{An \code{\link[=narrow_array]{narrow_array()}}.}
}
\value{
An \code{\link[=narrow_array]{narrow_array()}}
}
\description{
Create copies of Arrow vectors
}
\examples{
narrow_deep_copy(as_narrow_array(1:5))

}
