% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-ptype.R
\name{narrow_default_ptype}
\alias{narrow_default_ptype}
\title{Guess prototypes for Arrow schemas}
\usage{
narrow_default_ptype(schema)
}
\arguments{
\item{schema}{An \code{\link[=narrow_schema]{narrow_schema()}}}
}
\value{
A zero-length array prototype
}
\description{
Guess prototypes for Arrow schemas
}
\examples{
narrow_default_ptype(narrow_schema("n"))
narrow_default_ptype(
  narrow_schema("+s", children = list(narrow_schema("i", "colname")))
)

}
