% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointers.R
\name{narrow_pointer_is_valid}
\alias{narrow_pointer_is_valid}
\alias{narrow_pointer_release}
\alias{narrow_pointer_move}
\alias{narrow_pointer_export}
\alias{narrow_allocate_schema}
\alias{narrow_allocate_array_data}
\alias{narrow_allocate_array_stream}
\alias{narrow_pointer_addr_chr}
\alias{narrow_schema_from_pointer}
\alias{narrow_array_data_from_pointer}
\alias{narrow_array_stream_from_pointer}
\alias{narrow_from_pointer}
\title{Danger zone: low-level pointer operations}
\usage{
narrow_pointer_is_valid(ptr)

narrow_pointer_release(ptr)

narrow_pointer_move(ptr_src, ptr_dst)

narrow_pointer_export(ptr_src, ptr_dst)

narrow_allocate_schema()

narrow_allocate_array_data()

narrow_allocate_array_stream()

narrow_pointer_addr_chr(ptr)

narrow_schema_from_pointer(ptr)

narrow_array_data_from_pointer(ptr)

narrow_array_stream_from_pointer(ptr)

narrow_from_pointer(ptr, cls)
}
\arguments{
\item{ptr, ptr_src, ptr_dst}{An external pointer to a \code{\link[=narrow_schema]{narrow_schema()}},
a \code{\link[=narrow_array_data]{narrow_array_data()}}, or a \code{\link[=narrow_array_stream]{narrow_array_stream()}}.}

\item{cls}{One of "narrow_schema", "narrow_array_data", or
"narrow_array_stream".}
}
\description{
Danger zone: low-level pointer operations
}
