% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{narrow_vctr}
\alias{narrow_vctr}
\alias{as_narrow_vctr}
\alias{new_narrow_vctr}
\title{Create R vector wrappers around 'Arrow' arrays}
\usage{
narrow_vctr(array = narrow_array())

as_narrow_vctr(x, ...)

new_narrow_vctr(x = integer(), array = narrow_array())
}
\arguments{
\item{array}{A \code{\link[=narrow_array]{narrow_array()}}}

\item{x}{An object to convert to a \code{\link[=narrow_vctr]{narrow_vctr()}}}

\item{...}{Passed to \code{\link[=as_narrow_array]{as_narrow_array()}}}
}
\value{
An object of class 'narrow_vctr'
}
\description{
Create R vector wrappers around 'Arrow' arrays
}
\examples{
as_narrow_vctr(1:10)

}
