% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{parse_format}
\alias{parse_format}
\title{Parse Arrow format strings}
\usage{
parse_format(format)
}
\arguments{
\item{format}{A
\href{https://arrow.apache.org/docs/format/CDataInterface.html#data-type-description-format-strings}{format string}.}
}
\value{
A \code{\link[=list]{list()}} with elements:
\itemize{
\item \code{format}: The input string
\item \code{abbreviation}: The short name of the type
\item \code{spec}: The parsing pattern for the type
\item \code{description}: The human-readable name of the type
\item \code{args}: A \code{\link[=list]{list()}} of key/value pairs of type arguments
}
}
\description{
Parse Arrow format strings
}
\examples{
parse_format("i")
parse_format("w:128")
parse_format("+ud:1,2,3")

}
