% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{sparrow_array}
\alias{sparrow_array}
\alias{sparrow_array_validate}
\alias{as_sparrow_array}
\alias{as_sparrow_array.sparrow_array}
\title{Create arrow vectors}
\usage{
sparrow_array(
  schema = sparrow_schema("n"),
  array_data = sparrow_array_data(),
  validate = TRUE
)

sparrow_array_validate(x)

as_sparrow_array(x, ...)

\method{as_sparrow_array}{sparrow_array}(x, ...)
}
\arguments{
\item{schema}{An \code{\link[=sparrow_schema]{sparrow_schema()}}}

\item{array_data}{An \code{\link[=sparrow_array_data]{sparrow_array_data()}}}

\item{validate}{Use \code{FALSE} to skip validation}

\item{x}{An object to convert to an \code{\link[=sparrow_array]{sparrow_array()}}}

\item{...}{Passed to S3 methods}
}
\value{
An object of class "sparrow_array"
}
\description{
Create arrow vectors
}
\examples{
sparrow_array()

}
