% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-data.R
\name{sparrow_array_data}
\alias{sparrow_array_data}
\alias{sparrow_array_data_info}
\alias{as_sparrow_array_data}
\alias{as_sparrow_array_data.sparrow_array_data}
\alias{as_sparrow_buffer}
\alias{as_sparrow_buffer.default}
\title{Create Arrow arrays}
\usage{
sparrow_array_data(
  buffers = NULL,
  length = 0,
  null_count = -1,
  offset = 0,
  children = NULL,
  dictionary = NULL
)

sparrow_array_data_info(x, ...)

as_sparrow_array_data(x, ...)

\method{as_sparrow_array_data}{sparrow_array_data}(x, ...)

as_sparrow_buffer(x, ...)

\method{as_sparrow_buffer}{default}(x, ...)
}
\arguments{
\item{buffers}{A \code{\link[=list]{list()}} of R vectors whose memory serves as an underlying
buffer in the Arrow format.}

\item{length}{The logical size of the array}

\item{null_count}{The number of NULL values or -1 if this hasn't been
computed yet.}

\item{offset}{The number of elements to skip at the front of the array.}

\item{children}{Child vectors as a \code{\link[=list]{list()}} coerced by \code{\link[=as_sparrow_array_data]{as_sparrow_array_data()}}}

\item{dictionary}{Dictionary array for dictionary types.}

\item{x}{An object to convert to an sparrow_array}

\item{...}{Passed to S3 Methods}
}
\value{
An object of class "sparrow_array_data"
}
\description{
Create Arrow arrays
}
\examples{
sparrow_array_data(1:100, 100)

}
