% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-stream-function.R
\name{sparrow_array_stream_function}
\alias{sparrow_array_stream_function}
\title{Array streams from functions}
\usage{
sparrow_array_stream_function(schema, fun, validate = TRUE)
}
\arguments{
\item{schema}{A \code{\link[=sparrow_schema]{sparrow_schema()}}}

\item{fun}{An R function whose return value is converted to
a \code{\link[=sparrow_array]{sparrow_array()}} and passed to the caller.}

\item{validate}{Use \code{FALSE} to skip validation of the output of the function
against \code{schema}.}
}
\value{
An object of class "sparrow_array_stream"
}
\description{
Array streams from functions
}
