% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{sparrow_deep_copy}
\alias{sparrow_deep_copy}
\title{Create copies of Arrow vectors}
\usage{
sparrow_deep_copy(x)
}
\arguments{
\item{x}{An \code{\link[=sparrow_array]{sparrow_array()}}.}
}
\value{
An \code{\link[=sparrow_array]{sparrow_array()}}
}
\description{
Create copies of Arrow vectors
}
\examples{
sparrow_deep_copy(as_sparrow_array(1:5))

}
