% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-ptype.R
\name{sparrow_default_ptype}
\alias{sparrow_default_ptype}
\title{Guess prototypes for Arrow schemas}
\usage{
sparrow_default_ptype(schema)
}
\arguments{
\item{schema}{An \code{\link[=sparrow_schema]{sparrow_schema()}}}
}
\value{
A zero-length array prototype
}
\description{
Guess prototypes for Arrow schemas
}
\examples{
sparrow_default_ptype(sparrow_schema("n"))
sparrow_default_ptype(
  sparrow_schema("+s", children = list(sparrow_schema("i", "colname")))
)

}
