% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{sparrow_vctr}
\alias{sparrow_vctr}
\alias{as_sparrow_vctr}
\title{Create R vector wrappers around 'Arrow' arrays}
\usage{
sparrow_vctr(array = sparrow_array())

as_sparrow_vctr(x, ...)
}
\arguments{
\item{array}{A \code{\link[=sparrow_array]{sparrow_array()}}}

\item{x}{An object to convert to a \code{\link[=sparrow_vctr]{sparrow_vctr()}}}

\item{...}{Passed to \code{\link[=as_sparrow_array]{as_sparrow_array()}}}
}
\value{
An object of class 'sparrow_vctr'
}
\description{
Create R vector wrappers around 'Arrow' arrays
}
\examples{
as_sparrow_vctr(1:10)

}
