% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{arrow_schema}
\alias{arrow_schema}
\alias{arrow_schema_flags}
\alias{arrow_schema_copy}
\alias{as_arrow_schema}
\alias{as_arrow_schema.arrowvctrs_schema}
\alias{as_arrow_schema.character}
\title{Create Arrow Schema objects}
\usage{
arrow_schema(
  format,
  name = NULL,
  metadata = NULL,
  flags = arrow_schema_flags(),
  children = NULL,
  dictionary = NULL
)

arrow_schema_flags(
  dictionary_ordered = FALSE,
  nullable = FALSE,
  map_keys_sorted = FALSE
)

arrow_schema_copy(schema)

as_arrow_schema(x, ...)

\method{as_arrow_schema}{arrowvctrs_schema}(x, ...)

\method{as_arrow_schema}{character}(x, ...)
}
\arguments{
\item{format}{A
\href{https://arrow.apache.org/docs/format/CDataInterface.html#data-type-description-format-strings}{format string}.}

\item{name}{An optional name}

\item{metadata}{A named character vector of \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}}
containing key/value metadata.}

\item{flags}{Flags to set on input (see \code{\link[=arrow_schema_flags]{arrow_schema_flags()}})}

\item{children}{A \code{\link[=list]{list()}} of objects created using \code{\link[=arrow_schema]{arrow_schema()}}.}

\item{dictionary}{An \code{\link[=arrow_schema]{arrow_schema()}} if this is a dictionary type.}

\item{dictionary_ordered}{\code{TRUE} if the \code{dictionary} array is ordered}

\item{nullable}{\code{TRUE} if the type is semantically nullable}

\item{map_keys_sorted}{\code{TRUE} if the keys for a map have been sorted}

\item{schema}{A schema created with \code{\link[=arrow_schema]{arrow_schema()}}}

\item{x}{An object to convert to an \code{\link[=arrow_schema]{arrow_schema()}}}

\item{...}{Passed to S3 methods}
}
\value{
An external pointer with class 'arrowvctrs_schema'
}
\description{
Create Arrow Schema objects
}
