% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{arrow_vctr}
\alias{arrow_vctr}
\alias{arrow_vctr_validate}
\alias{as_arrow_vctr}
\alias{as_arrow_vctr.arrowvctrs_vctr}
\title{Create arrow vectors}
\usage{
arrow_vctr(schema = arrow_schema("n"), array = arrow_array_data())

arrow_vctr_validate(x)

as_arrow_vctr(x, ...)

\method{as_arrow_vctr}{arrowvctrs_vctr}(x, ...)
}
\arguments{
\item{schema}{An \code{\link[=arrow_schema]{arrow_schema()}}}

\item{array}{An \code{\link[=arrow_array_data]{arrow_array_data()}}}

\item{x}{An object to convert to an \code{\link[=arrow_vctr]{arrow_vctr()}}}

\item{...}{Passed to S3 methods}
}
\value{
An object of class "arrowvctrs_vctr"
}
\description{
Create arrow vectors
}
\examples{
arrow_vctr()

}
