% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{carrow_array}
\alias{carrow_array}
\alias{carrow_array_validate}
\alias{as_carrow_array}
\alias{as_carrow_array.arrowvctrs_vctr}
\title{Create arrow vectors}
\usage{
carrow_array(schema = arrow_schema("n"), array = arrow_array_data())

carrow_array_validate(x)

as_carrow_array(x, ...)

\method{as_carrow_array}{arrowvctrs_vctr}(x, ...)
}
\arguments{
\item{schema}{An \code{\link[=arrow_schema]{arrow_schema()}}}

\item{array}{An \code{\link[=arrow_array_data]{arrow_array_data()}}}

\item{x}{An object to convert to an \code{\link[=carrow_array]{carrow_array()}}}

\item{...}{Passed to S3 methods}
}
\value{
An object of class "arrowvctrs_vctr"
}
\description{
Create arrow vectors
}
\examples{
carrow_array()

}
