% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-from-vctr.R
\name{from_arrow_vctr}
\alias{from_arrow_vctr}
\alias{from_arrow_vctr.default}
\alias{from_arrow_vctr.NULL}
\alias{from_arrow_vctr.logical}
\alias{from_arrow_vctr.integer}
\alias{from_arrow_vctr.double}
\alias{from_arrow_vctr.character}
\alias{from_arrow_vctr.factor}
\alias{from_arrow_vctr.data.frame}
\title{Convert Arrow vectors to R objects}
\usage{
from_arrow_vctr(x, ptype, ...)

\method{from_arrow_vctr}{default}(x, ptype, ...)

\method{from_arrow_vctr}{`NULL`}(x, ptype, ...)

\method{from_arrow_vctr}{logical}(x, ptype, ...)

\method{from_arrow_vctr}{integer}(x, ptype, ...)

\method{from_arrow_vctr}{double}(x, ptype, ...)

\method{from_arrow_vctr}{character}(x, ptype, ...)

\method{from_arrow_vctr}{factor}(x, ptype, ...)

\method{from_arrow_vctr}{data.frame}(x, ptype, ...)
}
\arguments{
\item{x}{An \code{\link[=arrow_vctr]{arrow_vctr()}}}

\item{ptype}{An R object to use as a prototype}

\item{...}{Passed to S3 methods}
}
\value{
An object with the same class as \code{ptype}
}
\description{
Note that \code{\link[=from_arrow_vctr]{from_arrow_vctr()}} dispatches on \code{ptype}
}
\examples{
from_arrow_vctr(as_arrow_vctr(NULL), NULL)

}
