% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-arrow.R
\name{from_arrow_vctr.Array}
\alias{from_arrow_vctr.Array}
\alias{from_arrow_vctr.R6ClassGenerator}
\alias{as_arrow_schema.Type}
\alias{as_arrow_schema.Field}
\alias{as_arrow_schema.Schema}
\alias{as_arrow_vctr.Array}
\alias{as_arrow_vctr.RecordBatch}
\title{Convert to and from arrow package types}
\usage{
\method{from_arrow_vctr}{Array}(x, ptype, ...)

\method{from_arrow_vctr}{R6ClassGenerator}(x, ptype, ...)

\method{as_arrow_schema}{Type}(x, ...)

\method{as_arrow_schema}{Field}(x, ...)

\method{as_arrow_schema}{Schema}(x, ...)

\method{as_arrow_vctr}{Array}(x, ...)

\method{as_arrow_vctr}{RecordBatch}(x, ...)
}
\arguments{
\item{x}{An object to conver to or from 'arrow' package types}

\item{ptype}{An R object to use as a prototype}

\item{...}{Passed to S3 methods}
}
\description{
Convert to and from arrow package types
}
