% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-arrow.R
\name{from_carrow_array.Array}
\alias{from_carrow_array.Array}
\alias{from_carrow_array.R6ClassGenerator}
\alias{as_carrow_schema.DataType}
\alias{as_carrow_schema.Field}
\alias{as_carrow_schema.Schema}
\alias{as_carrow_array.Array}
\alias{as_carrow_array.RecordBatch}
\alias{carrow_array_stream_to_arrow}
\alias{as_carrow_array_stream.RecordBatchReader}
\alias{as_carrow_array_stream.RecordBatchFileReader}
\alias{as_carrow_array_stream.RecordBatchStreamReader}
\title{Convert to and from arrow package types}
\usage{
\method{from_carrow_array}{Array}(x, ptype, ...)

\method{from_carrow_array}{R6ClassGenerator}(x, ptype, ...)

\method{as_carrow_schema}{DataType}(x, ...)

\method{as_carrow_schema}{Field}(x, ...)

\method{as_carrow_schema}{Schema}(x, ...)

\method{as_carrow_array}{Array}(x, ...)

\method{as_carrow_array}{RecordBatch}(x, ...)

carrow_array_stream_to_arrow(x)

\method{as_carrow_array_stream}{RecordBatchReader}(x, ...)

\method{as_carrow_array_stream}{RecordBatchFileReader}(x, ...)

\method{as_carrow_array_stream}{RecordBatchStreamReader}(x, ...)
}
\arguments{
\item{x}{An object to convert to or from 'arrow' package types}

\item{ptype}{An R object to use as a prototype}

\item{...}{Passed to S3 methods}
}
\description{
Convert to and from arrow package types
}
